Move error types to their own file
